refactor(zimic-test-client): use `@zimic/http`
