+++ apps/zimic-test-client/package.json
@@ -20,6 +20,7 @@
     "@vitest/browser": "^3.0.5",
     "@vitest/coverage-istanbul": "^3.0.5",
     "@zimic/eslint-config-node": "workspace:*",
+    "@zimic/http": "workspace:*",
     "@zimic/lint-staged-config": "workspace:*",
     "@zimic/tsconfig": "workspace:*",
     "axios": "^1.7.9",
