+++ packages/zimic/src/interceptor/http/index.ts
@@ -1,5 +1,7 @@
 import HttpInterceptorNamespace from './namespace/HttpInterceptorNamespace';
 
+export { default as InvalidJSONError } from './interceptorWorker/errors/InvalidJSONError';
+export { default as InvalidFormDataError } from './interceptorWorker/errors/InvalidFormDataError';
 export { default as NotStartedHttpInterceptorError } from './interceptor/errors/NotStartedHttpInterceptorError';
 export { default as UnknownHttpInterceptorPlatformError } from './interceptor/errors/UnknownHttpInterceptorPlatformError';
 export { default as UnknownHttpInterceptorTypeError } from './interceptor/errors/UnknownHttpInterceptorTypeError';
